Major update for bricks. Added jquery.tmpl/tmplPlus. Created ccm_widget base class for jQuery widgets. This widget uses jQuery.metadata to gather properties from element attribute data-options-(widgetName). Created ccm_contextMenu jquery widget. Made AttributeKeyCategoryAvailableColumnSet return standard properties and all attributes (as I think is the expected behavior, rather than just extending DefaultColumnSet).
Item Selector:
Added create brick action, item context menus, and ajax retrieval of row data.
Created general get_item_properties tool for akci json.
TODO: Modify the search controller to accept json requests for this purpose or use REST service.

Item Search:
Big cleanup. Results are coming through the controller now, rather than recreating the controller and calling its methods from the elements.
TODO:
    Convert column config js to jquery widget and create a form helper.
    Add context menus and convert item action parameters to be like item selector.
    Convert additional search form attributes to use jQuery.tmpl.
BUGS: Sorting and paging don't appear to be working, even though the request variables are there. Not sure why.

Item Insert:
Added tool for creating brick items via dialog. Separated insert form into elements for reuse (ideally for reuse on item edit, too).
TODO: Duplicate functionality for Item Edit.

Item Edit:
Cleanup and prep for todos.
TODO:
   Modify/employ item insert elements for reuse. Duplicate tool dialog functionality for edit.

Item Delete:
Added permission checking before delete.

Bulk Properties:
Minor cleanup.
BUGS: Something is off on the saving. I reverted my changes to see if it was something I did, but it doesn't appear so. I think it might be related to whatever the GENERAL BUG below is.

----
GENERAL BUGS:
Seems like a brick type will just crap out after a little while. Not sure why/when this happens. I've seen different kinds of errors after little or no data change (all api), and then the errors just won't go away.

GENERAL TODO:
Create ccm_dialog to replace global override of jquery.ui.dialog. Current approach is too intrusive.

+++ web/concrete/tools/bricks/bulk_delete.php
@@ -1,44 +1,59 @@
 <?php defined('C5_EXECUTE') or die(_("Access Denied."));
 
-$searchInstance = isset($_REQUEST['searchInstance']) ? $_REQUEST['searchInstance'] : NULL;
+$baseId = isset($_REQUEST['baseId']) ? $_REQUEST['baseId'] : $_REQUEST['akCategoryHandle'];
+$wrapId = $baseId.'_bulk_delete_form';
+
+$json = Loader::helper('json');
+$form = Loader::helper('form');
+
+Loader::model('attribute_key_category_item_permission');
 
 if ($_REQUEST['task'] == 'delete') {
 	if (!ini_get('safe_mode')) {
 		@set_time_limit(0);
 	}
-	//$js = Loader::helper('json');
-	//$decoded = $js->decode($_REQUEST['json']);
+	$output = array('deleted'=>array());
 	$akc = AttributeKeyCategory::getByHandle($_REQUEST['akCategoryHandle']);
-	foreach($_REQUEST['akcID'] as $ID) {
-		$akci = $akc->getItemObject($ID);
-		// REALLY NEED TO CHECK PERMISSIONS HERE
-		$akci->delete(); 
+	foreach($_REQUEST['akciID'] as $ID) {
+		$akci = $akc->getItemObject($ID);		
+		$akcip = AttributeKeyCategoryItemPermission::get($akci);
+		if($akcip->canDelete()){
+			$output['deleted'][] = $akci->getID();
+			$akci->delete();		
+		}
 	}
+	echo $json->encode($output);
+	
 } else { ?>
-	<form id="<?php echo $searchInstance ?>_delete" method="post" action="<?php echo $_SERVER['PHP_SELF'].'?'.$_SERVER['QUERY_STRING'] ?>&task=delete">
-    	<?php //$form->hidden('akCategoryHandle', $_REQUEST['akCategoryHandle']) ?>
-        <?php //$form->hidden('searchInstance', $_REQUEST['akCategoryHandle']) ?>
-		<p>Are you sure you want to delete the selected items?</p>
+	<form id="<?php echo $wrapId ?>" method="post" action="<?php echo $_SERVER['PHP_SELF'] ?>">
+		<?php echo $form->hidden('task', 'delete') ?>
+    	<?php echo $form->hidden('akCategoryHandle', $_REQUEST['akCategoryHandle']) ?>
+        <?php echo $form->hidden('baseId', $baseId) ?>
     
 		<?php
-            //$js = Loader::helper('json');
-            //$decoded = $js->decode($_REQUEST['json']);
-            $akc = AttributeKeyCategory::getByHandle($_REQUEST['akCategoryHandle']);
-            //foreach($decoded->akcIDs as $ID) {
-                //$akci = $akc->getItemObject($ID);
-            //}
-            $ih = Loader::helper('concrete/interface');
             
-            print $ih->button_js(t('Yes'), '$(\'#'.$searchInstance.'_delete\').submit();');
-            print $ih->button_js(t('No'), 'jQuery.fn.dialog.closeTop()');
+            $akc = AttributeKeyCategory::getByHandle($_REQUEST['akCategoryHandle']);
+            foreach($_REQUEST['akciID'] as $ID) {
+                $akci = $akc->getItemObject($ID);
+				echo $form->hidden('akciID[]', $akci->getID());
+            }            
         ?>
+        
+        <p>Are you sure you want to delete the selected items?</p>
+        
+        <?php
+        
+        	$ih = Loader::helper('concrete/interface');            
+            print $ih->submit(t('Yes'));
+            print $ih->button_js(t('No'), 'jQuery.fn.dialog.closeTop()', 'LEFT');
+		?>
         
         <script type="text/javascript">
             $(function(){
-                $("#<?php echo $searchInstance ?>_delete").ajaxForm({
+                $("#<?php echo $wrapId ?>").ajaxForm({
                     success:function(){
-                        $.fn.dialog.closeTop();	
-                        $("#<?php echo $searchInstance ?>_form").submit();
+                        $.fn.dialog.closeTop();
+                        $("#<?php echo $baseId ?>_form").submit();
                     }
                 });
             });
